refactor: fix api-schema to schema dir
