+++ apps/web/src/services/alert-manager/alert/alert-detail/modules/AlertNote.vue
@@ -61,12 +61,13 @@ import { iso8601Formatter } from '@cloudforet/core-lib';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
-import type { TimeStamp } from '@/api-schema/common/model';
+import type { TimeStamp } from '@/schema/common/model';
 import { store } from '@/store';
 import { i18n } from '@/translations';
 
 import DeleteModal from '@/common/components/modals/DeleteModal.vue';
 import ErrorHandler from '@/common/composables/error/errorHandler';
+
 
 interface NoteModel {
     note_id: string;
