+++ apps/web/src/services/asset-inventory/cloud-service/cloud-service-detail/modules/CloudServiceDetail.vue
@@ -17,8 +17,8 @@ import { QueryHelper } from '@cloudforet/core-lib/query';
 import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 
-import type { ExportParameter } from '@/api-schema/common/api-verbs/export';
-import { QueryType } from '@/api-schema/common/api-verbs/export';
+import { QueryType } from '@/schema/common/api-verbs/export';
+import type { ExportParameter } from '@/schema/common/api-verbs/export';
 import { store } from '@/store';
 import { i18n } from '@/translations';
 
@@ -35,6 +35,7 @@ import { useQuerySearchPropsWithSearchSchema } from '@/common/composables/dynami
 import ErrorHandler from '@/common/composables/error/errorHandler';
 
 import { useCloudServiceDetailPageStore } from '@/services/asset-inventory/store/cloud-service-detail-page-store';
+
 
 interface Props {
     cloudServiceId: string;
