+++ apps/web/src/services/asset-inventory/service-account/type.ts
@@ -1,11 +1,13 @@
 import type { JsonSchema } from '@spaceone/design-system/types/inputs/forms/json-schema-form/type';
 
-import type { Tags, TimeStamp } from '@/api-schema/common/model';
+
+import type { Tags, TimeStamp } from '@/schema/common/model';
 
 import type { Tag } from '@/common/components/forms/tags-input-group/type';
 
 import type { ACCOUNT_TYPE } from '@/services/asset-inventory/service-account/config';
 import type { ProjectItemResp } from '@/services/project/type';
+
 
 const idField = 'provider';
