+++ apps/web/src/services/cost-explorer/components/BudgetMainList.vue
@@ -13,10 +13,6 @@ import { SpaceConnector } from '@cloudforet/core-lib/space-connector';
 import { ApiQueryHelper } from '@cloudforet/core-lib/space-connector/helper';
 import type { Query } from '@cloudforet/core-lib/space-connector/type';
 
-import type {
-    BudgetUsageAnalyzeResponse,
-    BudgetUsageAnalyzeResult,
-} from '@/api-schema/cost-analysis/budget-usage/api-verbs/analyze';
 import { store } from '@/store';
 
 import { FILE_NAME_PREFIX } from '@/lib/excel-export';
@@ -30,6 +26,11 @@ import {
     getBudgetUsageAnalyzeRequestQuery,
 } from '@/services/cost-explorer/helpers/budget-usage-analyze-api-query-helper';
 import type { Period } from '@/services/cost-explorer/types/cost-explorer-query-type';
+
+import type {
+    BudgetUsageAnalyzeResponse,
+    BudgetUsageAnalyzeResult,
+} from '@/schema/cost-analysis/budget-usage/api-verbs/analyze';
 
 
 interface Props {
