+++ apps/web/src/services/cost-explorer/components/BudgetMainListCard.vue
@@ -6,7 +6,6 @@ import {
     PDivider, PI,
 } from '@spaceone/design-system';
 
-import type { BudgetUsageAnalyzeResult } from '@/api-schema/cost-analysis/budget-usage/api-verbs/analyze';
 import { store } from '@/store';
 
 import type { ProjectGroupReferenceItem, ProjectGroupReferenceMap } from '@/store/modules/reference/project-group/type';
@@ -22,6 +21,8 @@ import { currencyMoneyFormatter } from '@/lib/helper/currency-helper';
 
 import BudgetMainUsageProgressBar from '@/services/cost-explorer/components/BudgetMainUsageProgressBar.vue';
 import { COST_EXPLORER_ROUTE } from '@/services/cost-explorer/routes/route-config';
+
+import type { BudgetUsageAnalyzeResult } from '@/schema/cost-analysis/budget-usage/api-verbs/analyze';
 
 
 interface Props {
