+++ apps/web/src/services/cost-explorer/helpers/budget-usage-analyze-api-query-helper.ts
@@ -2,10 +2,11 @@ import { isEmpty } from 'lodash';
 
 import type { Query } from '@cloudforet/core-lib/space-connector/type';
 
-import type { BudgetUsageAnalyzeResult } from '@/api-schema/cost-analysis/budget-usage/api-verbs/analyze';
 
 import { GRANULARITY } from '@/services/cost-explorer/constants/cost-explorer-constants';
 import type { Period } from '@/services/cost-explorer/types/cost-explorer-query-type';
+
+import type { BudgetUsageAnalyzeResult } from '@/schema/cost-analysis/budget-usage/api-verbs/analyze';
 
 
 type BudgetUsageAnalyzeRequestGroupBy = Partial<keyof BudgetUsageAnalyzeResult>;
